+++ SKIRTcore/DustSystemDensityCalculator.hpp
@@ -39,7 +39,7 @@ public:
         is designed for use as the body in a parallel loop; see the Parallel class. You must invoke
         this function for all indices in the range 0 thru \em numBodies-1 before calling the other
         functions in this class. */
-    void body(int n);
+    void body(size_t n);
 
     /** This function returns an estimate for the mean value of the
         difference \f$|\rho_g-\rho_t|\f$. */
